requirements.txt
streamlit
streamlit-mic-recorder
openai-whisper
python-dotenv
dateparser
apscheduler
watchdog